+++ src/test/java/nextstep/subway/member/MemberAcceptanceTest.java
@@ -1,19 +1,25 @@
 package nextstep.subway.member;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.member.dto.MemberRequest;
-import nextstep.subway.member.dto.MemberResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import static nextstep.subway.auth.acceptance.AuthAcceptanceTestMethod.로그인_요청;
-import static org.assertj.core.api.Assertions.assertThat;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.내정보_삭제_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.내정보_수정_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.내정보_조회_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.로그인_성공_이후_토큰;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_삭제_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_삭제됨;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_생성됨;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_생성을_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_정보_수정_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_정보_수정됨;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_정보_조회_요청;
+import static nextstep.subway.member.MemberAcceptanceTestMethod.회원_정보_조회됨;
 
 public class MemberAcceptanceTest extends AcceptanceTest {
     public static final String EMAIL = "email@email.com";
@@ -23,7 +29,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     public static final int AGE = 20;
     public static final int NEW_AGE = 21;
 
-    private ExtractableResponse<Response> 회원_생성_요청_응답;
+    private String 로그인_토큰;
 
     @BeforeEach
     void before() {
@@ -61,90 +67,23 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     void manageMyInfo() {
         //when: 로그인 요청
         ExtractableResponse<Response> 로그인_요청_응답 = 로그인_요청(EMAIL, PASSWORD);
-        String 로그인_토큰 = 로그인_성공_이후_토큰(로그인_요청_응답);
+        로그인_토큰 = 로그인_성공_이후_토큰(로그인_요청_응답);
 
-        //then: 내정보 조회
+        //then: 내 정보 조회
         ExtractableResponse<Response> 내정보_조회_요청_응답 = 내정보_조회_요청(로그인_토큰);
-        assertThat(내정보_조회_요청_응답.statusCode()).isEqualTo(HttpStatus.OK);
+        회원_정보_조회됨(내정보_조회_요청_응답, EMAIL, AGE);
 
+        //when: 내 정보 수정
+        ExtractableResponse<Response> 내정보_수정요청_응답 = 내정보_수정_요청(로그인_토큰, NEW_EMAIL, NEW_PASSWORD, NEW_AGE);
 
+        //then: 내 정보 수정됨
+        회원_정보_수정됨(내정보_수정요청_응답);
+
+        //when: 내 정보 삭제
+        ExtractableResponse<Response> 내정보_삭제_요청_응답 = 내정보_삭제_요청(로그인_토큰);
+
+        //then: 내 정보 삭제
+        회원_삭제됨(내정보_삭제_요청_응답);
     }
 
-    public static String 로그인_성공_이후_토큰(ExtractableResponse<Response> response) {
-        return response.jsonPath().getString("accessToken");
-    }
-
-    private ExtractableResponse<Response> 내정보_조회_요청(String token) {
-        return RestAssured.given().log().all().
-                auth().
-                oauth2(token).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().get("/members/me").
-                then().log().all().
-                extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age) {
-        MemberRequest memberRequest = new MemberRequest(email, password, age);
-
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberRequest)
-                .when().post("/members")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_정보_조회_요청(ExtractableResponse<Response> response) {
-        String uri = response.header("Location");
-
-        return RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get(uri)
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_정보_수정_요청(ExtractableResponse<Response> response, String email, String password, Integer age) {
-        String uri = response.header("Location");
-        MemberRequest memberRequest = new MemberRequest(email, password, age);
-
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberRequest)
-                .when().put(uri)
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_삭제_요청(ExtractableResponse<Response> response) {
-        String uri = response.header("Location");
-        return RestAssured
-                .given().log().all()
-                .when().delete(uri)
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 회원_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    public static void 회원_정보_조회됨(ExtractableResponse<Response> response, String email, int age) {
-        MemberResponse memberResponse = response.as(MemberResponse.class);
-        assertThat(memberResponse.getId()).isNotNull();
-        assertThat(memberResponse.getEmail()).isEqualTo(email);
-        assertThat(memberResponse.getAge()).isEqualTo(age);
-    }
-
-    public static void 회원_정보_수정됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 회원_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
 }
